Split internal/external flow helpers so that flows can diverge (#672)
* Split internal/external flow helpers so that flows can diverge

* Diverge internal/external return flow steps

* Move shared flow helpers to common file
