+++ src/modules/returns/lib/flow-helpers/steps.js
@@ -1,0 +1,39 @@
+const STEP_LICENCES = '/licences';
+const STEP_RETURNS = '/returns';
+const STEP_INTERNAL_ROUTING = '/return/internal';
+const STEP_LOG_RECEIPT = '/return/log-receipt';
+const STEP_RECEIPT_LOGGED = '/return/receipt-logged';
+const STEP_QUERY_LOGGED = '/return/query-logged';
+const STEP_START = '/return';
+const STEP_NIL_RETURN = '/return/nil-return';
+const STEP_METHOD = '/return/method';
+const STEP_METER_DETAILS = '/return/meter/details';
+const STEP_METER_UNITS = '/return/meter/units';
+const STEP_METER_READINGS = '/return/meter/readings';
+const STEP_UNITS = '/return/units';
+const STEP_SINGLE_TOTAL = '/return/single-total';
+const STEP_BASIS = '/return/basis';
+const STEP_QUANTITIES = '/return/quantities';
+const STEP_CONFIRM = '/return/confirm';
+const STEP_SUBMITTED = '/return/submitted';
+
+module.exports = {
+  STEP_LICENCES,
+  STEP_RETURNS,
+  STEP_INTERNAL_ROUTING,
+  STEP_LOG_RECEIPT,
+  STEP_RECEIPT_LOGGED,
+  STEP_START,
+  STEP_NIL_RETURN,
+  STEP_METHOD,
+  STEP_UNITS,
+  STEP_SINGLE_TOTAL,
+  STEP_BASIS,
+  STEP_QUANTITIES,
+  STEP_METER_DETAILS,
+  STEP_METER_UNITS,
+  STEP_METER_READINGS,
+  STEP_CONFIRM,
+  STEP_SUBMITTED,
+  STEP_QUERY_LOGGED
+};
